@@ -23,7 +23,7 @@ import com.github.elopteryx.upload.UploadContext;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import javax.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletRequest;
 
 /**
  * Default implementation of {@link UploadContext}.
